services/foundations_extraction/requirements.txt
amazon-textract-caller==0.2.3
amazon-textract-response-parser==1.0.2
amazon-textract-textractor==1.7.11
annotated-types==0.6.0
anyio==4.3.0
boto3==1.34.108
botocore==1.34.108
certifi==2024.07.04
click==8.1.7
dnspython==2.6.1
dyntastic==0.15.0
editdistance==0.8.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
orjson==3.10.3
packaging==24.0
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3transfer==0.10.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tabulate==0.9.0
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==1.26.19
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
XlsxWriter==3.2.0
